% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padtwins.R
\name{padtwins.function}
\alias{padtwins.function}
\title{Padtwins Method for Function Input}
\usage{
\method{padtwins}{`function`}(
  A,
  generate_P = function(A) matrix(stats::rbinom(n = nrow(A) * ncol(A), size = 1, prob =
    0.5), nrow = nrow(A)),
  generate_D = function(A) c(stats::rgeom(n = nrow(A), prob = 0.5)),
  ...
)
}
\arguments{
\item{A}{A function to generate matrix 'A'.}

\item{generate_P}{A function to generate matrix 'P' based on the dimensions of 'A'. Defaults to a function that generates a matrix with binomially distributed numbers.}

\item{generate_D}{A function to generate a diagonal matrix 'D'. Defaults to a function that generates a vector with geometrically distributed numbers.}

\item{...}{Additional arguments to be passed to specific methods.}
}
\value{
An object of class 'padtwins' containing the results of the matrix transformation.
}
\description{
This method of the padtwins function handles the scenario where the input 'A' is a function that generates a matrix. It allows dynamic generation of matrix 'A' and performs specific matrix transformations.
}
