% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padtwins.R
\name{padtwins.matrix}
\alias{padtwins.matrix}
\title{Padtwins Method for Matrix Input}
\usage{
\method{padtwins}{matrix}(
  A,
  generate_P = function(A) matrix(stats::rbinom(n = nrow(A) * ncol(A), size = 1, prob =
    0.5), nrow = nrow(A)),
  generate_D = function(A) c(stats::rgeom(n = nrow(A), prob = 0.5)),
  ...
)
}
\arguments{
\item{A}{A predefined square matrix.}

\item{generate_P}{A function to generate matrix 'P'. Defaults to a function that generates a matrix with binomially distributed numbers.}

\item{generate_D}{A function to generate a diagonal matrix 'D'. Defaults to a function that generates a vector with geometrically distributed numbers.}

\item{...}{Additional arguments to be passed to specific methods.}
}
\value{
An object of class 'padtwins' containing the results of the matrix transformation.
}
\description{
This method of the padtwins function handles the scenario where the input 'A' is a predefined matrix. It performs specific matrix transformations based on matrices 'P' and 'D' generated using provided anonymous functions or default random generators.
}
