% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padtwins.R
\name{padtwins}
\alias{padtwins}
\title{Padtwins Function}
\usage{
padtwins(
  A,
  generate_P = function(A) matrix(stats::rbinom(n = nrow(A) * ncol(A), size = 1, prob =
    0.5), nrow = nrow(A)),
  generate_D = function(A) c(stats::rgeom(n = nrow(A), prob = 0.5)),
  ...
)
}
\arguments{
\item{A}{A matrix or a function to generate a matrix. If a matrix is provided, it should be square. If a function is provided, it should return a square matrix.}

\item{generate_P}{A function to generate matrix 'P' based on the dimensions of 'A'. Defaults to a function that generates a matrix with binomially distributed numbers.}

\item{generate_D}{A function to generate a diagonal matrix 'D'. Defaults to a function that generates a vector with geometrically distributed numbers.}

\item{...}{Additional arguments to be passed to specific methods.}
}
\value{
An object of class 'padtwins' containing the results of the matrix transformation.
}
\description{
This is a generic function for the padtwins package. It takes either a matrix or a function to generate a matrix and computes a specific matrix transformation.
}
\examples{
# Using a predefined matrix
set.seed(102)
A <- matrix(c(1, 2, -2, 1), nrow = 2, ncol = 2)
result <- padtwins(A)

# Using a function to generate matrix A
set.seed(110)
generate_A <- function() matrix(rhyper(4, 4, 5, 3), nrow = 2, ncol = 2)
padt <- padtwins(generate_A)
}
