% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_sim.R
\name{plot.quick_sim_result}
\alias{plot.quick_sim_result}
\title{Plot Method for Quick Simulation Results}
\usage{
\method{plot}{quick_sim_result}(x, ...)
}
\arguments{
\item{x}{An object of class 'quick_sim_result'.}

\item{...}{Additional arguments affecting the plot produced.}
}
\value{
A bar plot of the simulation results.
}
\description{
Plots the quick simulation results, showing the proportion of successful runs, invertibility of matrix P, and success in diagonalizing matrix A in a bar chart.
}
\examples{
set.seed(102)
A <- matrix(c(1, 2, -2, 1), nrow = 2, ncol = 2)
sim_result <- quick_sim(function() padtwins(A), iterations = 100)
plot(sim_result)
}
