% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_sim.R
\name{quick_sim}
\alias{quick_sim}
\title{Quick Simulation for Padtwins Function}
\usage{
quick_sim(padtwins_func, iterations = 100)
}
\arguments{
\item{padtwins_func}{A padtwins function to be simulated.}

\item{iterations}{Number of iterations for the simulation (default: 100).}
}
\value{
An object of class 'quick_sim_result' containing three logical vectors: run_without_error, p_invertible, and a_diagonalized.
}
\description{
Runs a quick simulation on a padtwins function for a given number of iterations. It records whether the function runs without error, whether matrix P is invertible, and whether matrix A is successfully diagonalized.
}
\examples{
set.seed(102)
A <- matrix(c(1, 2, -2, 1), nrow = 2, ncol = 2)
sim_result <- quick_sim(function() padtwins(A), iterations = 100)
summary(sim_result)
plot(sim_result)
}
