% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_sim.R
\name{summary.quick_sim_result}
\alias{summary.quick_sim_result}
\title{Summary Method for Quick Simulation Results}
\usage{
\method{summary}{quick_sim_result}(object, ...)
}
\arguments{
\item{object}{An object of class 'quick_sim_result'.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
A printed summary of simulation results.
}
\description{
Generates a summary of the quick simulation results, showing the proportion of successful runs, invertibility of matrix P, and success in diagonalizing matrix A.
}
\examples{
set.seed(102)
A <- matrix(c(1, 2, -2, 1), nrow = 2, ncol = 2)
sim_result <- quick_sim(function() padtwins(A), iterations = 100)
summary(sim_result)
}
